refactor: init stdio/http servers without start
do not start the server after creating it, instead expose a method for it

also, the converse-mcp does not need to expose the jrpc2 types
